+++ R/coord-.R
@@ -184,10 +184,11 @@ Coord <- ggproto("Coord",
   # Will generally have to return FALSE for coordinate systems that enforce a fixed aspect ratio.
   is_free = function() FALSE,
 
-  setup_params = function(data) {
+  setup_params = function(self, data) {
     list(
       guide_default = guide_axis(),
-      guide_missing = guide_none()
+      guide_missing = guide_none(),
+      expand = parse_coord_expand(self$expand %||% TRUE)
     )
   },
 
@@ -241,6 +242,26 @@ render_axis <- function(panel_params, axis, scale, position, theme) {
   } else {
     zeroGrob()
   }
+}
+
+# Elaborates an 'expand' argument for every side (top, right, bottom or left)
+parse_coord_expand <- function(expand) {
+  check_logical(expand)
+  if (anyNA(expand)) {
+    cli::cli_abort("{.arg expand} cannot contain missing values.")
+  }
+
+  if (!is_named(expand)) {
+    return(rep_len(expand, 4))
+  }
+
+  # Match by top/right/bottom/left
+  out <- rep(TRUE, 4)
+  i <- match(names(expand), .trbl)
+  if (sum(!is.na(i)) > 0) {
+    out[i] <- unname(expand)[!is.na(i)]
+  }
+  out
 }
 
 # Utility function to check coord limits